IJ-34 Split the wba_self_permissions functionality into its own controller
